Mapeando o ORM das novas tabelas e relacionamentos que diz respeito ao Template Avaliacao Topico Pergunta.
