+++ src/main/java/com/infnet/avaliacao/entity/TemplateAvaliacaoTopicoPergunta.java
@@ -1,0 +1,48 @@
+package com.infnet.avaliacao.entity;
+
+import javax.persistence.*;
+import java.io.Serializable;
+
+@Entity
+@Table(name = "template_avaliacao_topico_pergunta")
+public class TemplateAvaliacaoTopicoPergunta implements Serializable {
+
+    private static final long serialVersionUID = 1L;
+
+    @OneToOne
+    @JoinColumn(name = "id_template_avaliacao", nullable = false)
+    private TemplateAvaliacao templateAvaliacao;
+
+    @OneToOne
+    @JoinColumn(name = "id_template_topico", nullable = false)
+    private TemplateTopico templateTopico;
+
+    @ManyToOne(cascade = CascadeType.ALL)
+    @JoinColumn(name = "id_template_pergunta", nullable = false)
+    private TemplatePergunta templatePergunta;
+
+    public TemplateAvaliacao getTemplateAvaliacao() {
+        return templateAvaliacao;
+    }
+
+    public void setTemplateAvaliacao(TemplateAvaliacao templateAvaliacao) {
+        this.templateAvaliacao = templateAvaliacao;
+    }
+
+    public TemplateTopico getTemplateTopico() {
+        return templateTopico;
+    }
+
+    public void setTemplateTopico(TemplateTopico templateTopico) {
+        this.templateTopico = templateTopico;
+    }
+
+    public TemplatePergunta getTemplatePergunta() {
+        return templatePergunta;
+    }
+
+    public void setTemplatePergunta(TemplatePergunta templatePergunta) {
+        this.templatePergunta = templatePergunta;
+    }
+
+}
